cleanup and new version
