+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [v0.0.7]
+
 - Allow for `cdo_options` and move from `extra` to `cdo_extra` (#31)
 - Refactor the logging (#31)
