+++ smmregrid/__init__.py
@@ -4,4 +4,4 @@ from .regrid import Regridder, regrid
 from .cdo_weights import cdo_generate_weights
 # from .util import find_vert_coords
 
-__version__ = '0.0.6'
+__version__ = '0.0.7'
